requirements.txt
pyTelegramBotAPI==4.13.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.8.5
